restructure package (#11)
* Features/GDEV-64

* Added .editorconfig and .gitattributes to normalise the code produced.

* Minor text change to .editorconfig.

* Removed forcing tab indentation as may cause issues with YAML no-tab indentation. We could set this up more granularly later.

* restructure package
