+++ .editorconfig
@@ -1,0 +1,15 @@
+# Top-most EditorConfig file.
+root = true
+
+# For all files ...
+[*]
+# Unix-style LF newlines.
+end_of_line = lf
+# Force a newline at the end of the file.
+insert_final_newline = true
+# Removed as may cause issues with YAML no-tab indentation. We could set this up more granularly later.
+# # Indents are tabs and have a size of 4 spaces.
+# indent_style = tab
+# indent_size = 4
+# Use UTF-8.
+charset = utf-8
