+++ sandbox_notification/__main__.py
@@ -1,0 +1,9 @@
+"""Main entrypoint for the service"""
+
+from .channels import subscribe
+
+def run():
+    subscribe()
+
+if __name__ == "__main__":
+    run()
